Adventure Missions progress - Viewing the adventure missions completed in a card is now possible.

+++ app/src/main/java/com/github/nacabaro/vbhelper/daos/CardProgressDao.kt
@@ -13,5 +13,5 @@ interface CardProgressDao {
     @Query(
         "SELECT currentStage FROM CardProgress WHERE cardId = :cardId"
     )
-    fun getCardProgress(cardId: Int): Int
+    fun getCardProgress(cardId: Long): Int
 }
